test: fix UDP conn ID generation test by mocking current time
The current implementation of `get_connection_id` seems to be OK.
I've added some tests. I needed to mock the current time becuase it's
used in the ID generation.

For now, I only neded to mock the current time (now). Not the clock
service.

I have also changed the signature of the `handle_connect` method.
Now it panics if we cannot get the system time. In the previous behavior
it was returning always a hardcoded connection ID:

ConnectionId(0x7FFFFFFFFFFFFFFF)