refactor: simplify trait ConnectionIdIssuer
Since we are not handling the error and other implementations could have
different errors I prefer to keep it simply.

The original idea was allowing the implementaions to raise concrete
errors messages. If we need to handle concrete error in the future for a
given implementaion we could add a new method "validate_connection_id".