feat: do not send response when connection id is not valid
BREAKING CHANGE: Following the BEP 15 (https://www.bittorrent.org/beps/bep_0015.html) we do not send amny response when the connection id is not valid in the UDP tracker.

In the previous behavior we were sending an error response with a
description.

We want to avoid involuntary smoke testing by malicious actors.

https://github.com/torrust/torrust-tracker/pull/60#issuecomment-1239191142