fix: weak tests
As I explained here:

https://github.com/torrust/torrust-tracker/pull/103#issue-1418647900

The mock for the trait TrackerStatisticsEventSender did not work
completely weel becuase it only checked for the rigth type of the
triggered event but It did not check if the event was sent.

I finally used a mockinf library becuase I do not know how to mock trait
that uses a mutable reference to 'self'. I need to store wether the
event was sent or not and I do not know how to do that without changing
the function signature making it mutable.