feat(api): the new Axum api uses the URL prefix /api too.
Initially we were using a ddifferent URl to avoid conflicts with the
Warp implementation but sicne we are using different ports that is not
a problem anymore. This change simplifies switching to the new Axum API,
since we only have to start using the new implementation in the
port set in the configuration (1212), instead of the temporarily created
port for the Axum implementation (1313).