fix(api): [#143] do not fail trying to remove a whitelisted torrent twice
Previous behavior:

When you try to remove a non exisinting whitelisted torrent the response
is 500 error.

New bahavior:

The enpoints checks if the torrent is included in the whitelist. If it
does not, then it ignores the reqeust returning a 200 code.

It should return a 204 or 404 but the current API only uses these codes:
200, 400, 405, 500. In the new API version we are planning to refctor
all endpoints.