refactor(http): [#159] rename struct in announce responses to follow new scrape conventions
Deserialization from bencoded bytes for announce and scrape request is
done in two phases. First using `serde_bencode::from_bytes` and later
with a custom parser. The reason is the `serde_bencode` crate does not
allow direct deserialization for the strcuts we need.

The strcut resulting from the first deserialization done by
`serde_bencode` is the `DeserializedCompact` and the second one just
`Compact`. So the prefix `Deserialized` is used when the bytes in the
reponse body are converted into a struct.