feat: [#187] add cargo dependency: bip_bencode
It will be used to build the bencoded compact annouce response in the
HTTP tracker. We are currently writting directly bytes into a byte
buffer, but Bencode specification imposes some restrictions like:

The keys in a dictionary must me alphabetically orderded. We are not
doing that in the current implementation.