feat: [#508] add health check enpoint to HTTP tracker
http://localhost:7070/health_check

And call the endpoint from the general application health check
endpoint:

http://localhost:1313/health_check

Do not call the endpoint if:

- The tracker is disabled.
- The tracker configuration uses port 0 only knwon after starting the
  socket.

todo: call the enpoint also when the port is 0 in the configuration. THe
service can return back to the main app the port assiged by the OS. And
the app can pass that port to the global app health check handler.