fix: add timeouts to UdpClient operations
The generic UDP client does not have timeouts. When the server is down
it waits forever for responses. This client has been using only for
testing where the server was always up, but now it's using in production
code by the Tracker Checker.

For the time being, I keep the panicking behavior of the UdpClient when
something is wrong. However we should return an error when the operation
times out.

For the Tracker Checker, it means that when the checker can't connect to
a UDP server the checker is going to panic after 5 seconds. That is not
the intended behavior for the checker. It should always return a full
reprot of all checks. In order to implement that behavior we need to
change the UdpClient to return errors. Since that's a bug refactor I
open a new issue.