dev: fix udp ring-buffer not looping
My previous version would be limited to a single thread, as `push_overwrite` would keep on returning the last element when the ring-buffer was full.

Now the ring-buffer is pre-filled and is looped over with a mutating iterator.

New handles are progressively swapped-in when the old entries are finished.

Note: I think that this implementation can be replaced with a standard vector with the same effect.