feat: enable overwrite Configuration values using env vars
Enable Figment ability to overwrite all config options with env vars.

We are currently overwriting only this value:

```toml
[http_api.access_tokens]
admin = "MyAccessToken"
```

With the env var `TORRUST_TRACKER_API_ADMIN_TOKEN`. The name we gave to
the env var does nto follow Figment convention which is `TORRUST_TRACKER_HTTP_API.ACCESS_TOKENS.ADMIN`.

We have to keep both options until we remove the old one in the rest of
the code.