tests: add test for configuration with deprecated env var name
Until it's updated in this repor and the Index repo you can overwrite
the admin token using the new env var name and the old one (deprecated):

- New: `TORRUST_TRACKER__HTTP_API__ACCESS_TOKENS__ADMIN`
- Old (deprecated): `TORRUST_TRACKER_API_ADMIN_TOKEN`

THe new one uses exactly the strcuture and atribute  names in the toml
file, using `__` as the separator for levels.

We can remove the test when we remove the deprecated name.