feat: [#870] add privacy methods to the TrackerMode
The tracker mode can be:

- Public (Non-whitelisted)
- Listed (Whitelisted)
- Private (Non-whitelisted)
- PrivateListed (Whitelisted)

They should have been two different flags (in my opinion):

- Visibility: public or private
- Whitelisted: true or false

So we would have the same four convinations:

- Not whitelisted:
 - Public
 - Private
- Whitelisted
 - Public
 - Private

That's a pending refactor. For this commits, the goal is just to align
this enum with what we added to the Index so we can use this enum in the
Index via the primmitives crate.