refactor: UDP tracker active request buffer
This refactoring tries to simplify the UDP active requests buffer by decoupling, making a place for a new request when the buffer is full and cleaning all finished tasks.

The idea is just to:

- First. If the buffer is full, just remove one item.
- After that, we can clean all finished tasks.

It was discarded because of the regression in performance.

I guess the problem is that, on average, we cannot guarantee that the eldest task will be the first to finish, so we have to remove more unfished tasks when there are finished tasks in the buffer.