feat: [#979] permanent keys
This commit adds a new feature. It allow creating permanent keys (keys
that do not expire).

THis is an example for making a request to the endpoint using curl:

```console
curl -X POST http://localhost:1212/api/v1/keys?token=MyAccessToken \
     -H "Content-Type: application/json" \
     -d '{
     	   "key": null,
           "seconds_valid": null
         }'
```

NOTICE: both the `key` and the `seconds_valid` fields can be null.

- If `key` is `null` a new random key will be generated. You can use an
  string with a pre-generated key like `Xc1L4PbQJSFGlrgSRZl8wxSFAuMa2110`. That will allow users to migrate to the Torrust Tracker wihtout forcing the users to re-start downloading/seeding with new keys.
- If `seconds_valid` is `null` the key will be permanent. Otherwise it
  will expire after the seconds specified in this value.