feat: [#978] add semantic validation for configuration
The section [core.provate_mode] can be inlcuded in the configration TOML
file only if the tracker is running in private mode (`private = true`).

This commits adds that validation and makes it possible to add more
semantic validations in the future.

Semantic validations are validations that depend on more than one value.
Like the one added, it could be any other incompatible combination.