refactor: [#1316] move health check API integration tests to pkg
It moved the integration tests for the Health Check API from the main
lib to the `axum-health-check-api-server`.

Some tests have not been moved becuase they depend on other server
pakages. They basically use the test "environments" from other servers which
are not publicly exposed yet. They are only used in integration tests.

We will move those environments to the corresponding server so they can
be used in other packages to run the servers.