refactor: [#1264] return wether swarm stats have change or not after upserting a peer
When you upsert a new peer (announce requests) the swarm stats migth
change if the peer announced that has completed the download. If that
case we return `true` given the caller the opportunity to know if stats
have changed.

In the current implementation we get stats before and after upserintg
the peer, and we compare them to check if they have changed. However
stats migth have change due to a parallel (race condition) announce from
another peer.

This is the only way to know exactly if this announce request has
altered the stats (incremented the number of downloads).