fix: [#1264] number of downloads preset when torrent is persisted
This fixed a bug: the number of donwloads for a torrent is not loaded
from the database when the torrent is added to the in-memory torrent
repository. It should be done when stats are enabled with the
configuration option:

persistent_torrent_completed_stat = true

The patch is applied only to the torrent repository implementation we
are using in prodcution (`CrossbeamSkipList<EntryMutexStd>`).

The other implementations have this comment:

```
// todo: load persistent torrent data if provided
```

It was not implemented for the others becuase I'm considering taking the
counter (for the number of downloads) out of the in-memory repository.
And increase it by suing events triggered from the core tracker.

I will open a new issue for that. If that's implemented we will need to
remove this patch for the repository, meaning reverting cahnges in this
commit.