refactor: [#1388] change channel in HTTP core from mpsc to broadcast
Stats events were introduced to collect tracker metrics. We only have global metrics (aggregate metrics for all UDP and HTTP trackers).

This will change in the future. We will have:

- Segregated metrics: one listeners per tracker (per socket).
- Generic events: there could be other event consumers. Events will be decoupled from stats.

This change allows multiple receivers in the channel.

For now, we one use one listener but with this change will be easy to add more.