test: [#1407] add test for global metrics with 2 http trackers
A new integration test that checks that the global metrics are udapted
when you run 2 HTTP trackers.

Ony one metric is checked in this test.

It uses fixed port that migth conflict with other running instances in
the future. We should use a random free port if we run more integration tests like this in the future.