fix: [#1477] shutdown event listeners on CRTL+c signal
This fixes the problem of adding the jobs for event listeners in the
main app JoinHandle vector. It was not possible to add the handles for
those tokio tasks becuase in the main app we wait for all jobs and those
jobs never end.

```rust
async fn main() {
    let (_app_container, jobs) = app::run().await;

    // handle the signals
    tokio::select! {
        _ = tokio::signal::ctrl_c() => {
            tracing::info!("Torrust tracker shutting down ...");

            // Await for all jobs to shutdown
            futures::future::join_all(jobs).await;

            tracing::info!("Torrust tracker successfully shutdown.");
        }
    }
}
```

Now, we can wait for them becuase they listen for the halt signal.

We will implement the shutdown in a different way in a new PR. See

https://github.com/torrust/torrust-tracker/issues/1405

Instead of listen to the CRTL+c signal the main app will send a "stop"
event to the listeners. The final goal it only the main app listen for
this external signal and it propagates the shutdown in cascade via
normal internal messages or channels.