fix: [#1510] disable torrent stats importation at start
When the tracker starts, if stats persistence is enabled, all torrents that have ever been downloaded are loaded into memory (`Swarms` type) with their download counter.

That's the current way to count all downloads and expose that metric.

However, it does not work with **millions of torrents** (like in the tracker demo) becuase:

- It's too slow.
- It consumes too much memory (all torrents that have ever been downloaded have to be loaded).

A new solution is needed to keep that metric, but in the meantime, this disables that feature, producing these effects:

- Non-accurate value for downloads when the tracker is restarted.
- Increasing indefinitely the number of torrents in memory even if the "remove peerless torrents" policy is enabled (becuase this feature overrides that policy and peerless torrents are kept in memory).