refactor: [#1358] rename metric
From `TORRENT_REPOSITORY_RUNTIME_TORRENTS_DOWNLOADS_TOTAL` to `TORRENT_REPOSITORY_TORRENTS_DOWNLOADS_TOTAL`.

None of the metrics in the `torrent-repositry` package will be
persisted. We can use the `persitent` sufix for metrics in other
packages to avoid conflicts.

It's planned to use the same metric in the `tracker-core` package but
with the historial persited value.