feat: [#1523] add new metric: number of inactive peers
The metric is added to the `torrent-repository` package.

The metric in Prometheus format:

```
torrent_repository_peers_inactive_total{} 0
```

It was not included as a new label in the number of peers because it can't be calculated from current events.

New inactivity events could have been added but the solution was much
more complex than this and having two metrics counting peers is not so
bad.

The discarded alternative was addinga new label por satte (`active`,
`inactive`).