fix: [#1543] Remove peerless torrents when it's enabled in the tracker policy
There were not being removed when stats was enabled becuase the tracker
was counting downloads only from the active swarms.

Now the API exposed metric (global downldoads) is not taken from the
in-memory data structrure unless stats persistence is disabled. In that
case, the global total would be per session (since the tracker started),
and reset when the tracker restarts.