fix: [#1514] bug. Don't allow merge metric collections with the same metrin name
It was not possible to merge counters or gauges if the metric names was
duplicate but possible when the metric name was duplciate across types.

For example, the target collection (the one is mutated) contains a counter with a name that is being used in the source collection (the wan we get metric from) for a gauge metric.